# viicasa-backend/docs/openapi.demo.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "VIICASA API · Fase 1",
    "version": "0.1.0",
    "description": "Backend independiente. Importes en centavos. Tokens Bearer distintos para visitantes y administradores."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "guestAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token obtenido de /v1/guest-sessions"
      }
    },
...
  },
  "paths": {
...
    "/v1/cart": {
      "get": {
        "tags": [
          "Carrito"
        ],
        "security": [
          {
            "guestAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/cart/items/{id}": {
      "put": {
        "tags": [
          "Carrito"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "quantity": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 99
                  }
                },
                "required": [
                  "quantity"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "guestAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "delete": {
        "tags": [
          "Carrito"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "guestAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/bookings/quote": {
      "post": {
        "tags": [
          "Reservaciones"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "property_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "check_in": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "check_out": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "guests": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100
                  }
                },
                "required": [
                  "property_id",
                  "check_in",
                  "check_out",
                  "guests"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/bookings": {
      "post": {
        "tags": [
          "Reservaciones"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "property_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "check_in": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "check_out": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "guests": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100
                  },
                  "customer": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 120
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 254
                      },
                      "phone": {
                        "type": "string",
                        "minLength": 5,
                        "maxLength": 30
                      },
                      "consent": {
                        "enum": [
                          true
                        ]
                      }
                    },
                    "required": [
                      "name",
                      "email",
                      "phone",
                      "consent"
                    ]
                  },
                  "pay": {
                    "type": "string",
                    "enum": [
                      "full",
                      "deposit"
                    ]
                  }
                },
                "required": [
                  "property_id",
                  "check_in",
                  "check_out",
                  "guests",
                  "customer",
                  "pay"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{8,100}$"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "guestAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/orders": {
      "post": {
        "tags": [
          "Pedidos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "customer": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "name": {
                        "type": "string",
                        "minLength": 2,
                        "maxLength": 120
                      },
                      "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 254
                      },
                      "phone": {
                        "type": "string",
                        "minLength": 5,
                        "maxLength": 30
                      },
                      "consent": {
                        "enum": [
                          true
                        ]
                      }
                    },
                    "required": [
                      "name",
                      "email",
                      "phone",
                      "consent"
                    ]
                  },
                  "delivery": {
                    "type": "string",
                    "enum": [
                      "pickup",
                      "shipping"
                    ]
                  },
                  "shipping_address": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "street": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 300
                      },
                      "city": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                      },
                      "state": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                      },
                      "postal_code": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 20
                      },
                      "country": {
                        "type": "string",
                        "pattern": "^[A-Z]{2}$"
                      }
                    },
                    "required": [
                      "street",
                      "city",
                      "state",
                      "postal_code",
                      "country"
                    ]
                  },
                  "note": {
                    "type": "string",
                    "minLength": 0,
                    "maxLength": 1000
                  }
                },
                "required": [
                  "customer",
                  "delivery"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{8,100}$"
            },
            "in": "header",
            "name": "idempotency-key",
            "required": true
          }
        ],
        "security": [
          {
            "guestAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/checkouts/{id}": {
      "get": {
        "tags": [
          "Operaciones"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "guestAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/checkouts/{id}/cancel": {
      "post": {
        "tags": [
          "Operaciones"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "guestAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/checkouts/{id}/payment": {
      "post": {
        "tags": [
          "Pagos"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "guestAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/payments/{id}/simulate": {
      "post": {
        "tags": [
          "Pagos demo"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "event_id": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 100
                  },
                  "outcome": {
                    "type": "string",
                    "enum": [
                      "paid",
                      "failed",
                      "expired"
                    ]
                  }
                },
                "required": [
                  "event_id",
                  "outcome"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/checkouts": {
      "get": {
        "tags": [
          "Operaciones"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000
            },
            "in": "query",
            "name": "offset",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "order",
                "booking"
              ]
            },
            "in": "query",
            "name": "kind",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "confirmed",
                "cancelled",
                "expired",
                "payment_review"
              ]
            },
            "in": "query",
            "name": "status",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/checkouts/{id}": {
      "get": {
        "tags": [
          "Operaciones"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/checkouts/{id}/cancel": {
      "post": {
        "tags": [
          "Operaciones"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/summary": {
      "get": {
        "tags": [
          "Resumen"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/mail-status": {
      "get": {
        "tags": [
          "Operación técnica"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/admin/audit": {
      "get": {
        "tags": [
          "Operación técnica"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 100000
            },
            "in": "query",
            "name": "offset",
            "required": false
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  }
}